Update rtx-venv.fish
Adds an error message for the case when a configuration file exists.

+++ rtx-venv.fish
@@ -1,3 +1,11 @@
 function rtx-venv -d "Configures a local RTX Python with a virtual environment"
+    if test -e .rtx.toml
+        echo -s \
+            (set_color $fish_color_error) \
+            'File .rtx.toml already exists; Better to edit it manually' \
+            (set_color normal)
+        return
+    end
+
     echo -ne '[tools]\npython = {version="latest", virtualenv=".venv"}\n' > .rtx.toml
 end
